pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-research-mcp"
version = "2.0.0"
description = "量化因子加工厂 - 因子的系统性提取、文档化、分析和策略回测平台"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Factor KB Team"}
]
keywords = ["quant", "factor", "trading", "knowledge-base", "data-hub", "strategy-hub", "backtest", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # 回测引擎
    "xbx-py11",
    # Web 框架
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",  # 文件上传支持
    # 网络请求
    "aiohttp>=3.9.0",
    # MCP SDK
    "mcp>=1.0.0",
    "sse-starlette>=2.0.0",
    # 配置
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # 日志
    "structlog>=24.0.0",
    # 数据库
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    # 数据库 ORM
    "sqlalchemy>=2.0.45",
    # LLM (LangChain)
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    # Research Hub - PDF 解析
    "pymupdf>=1.24.0",
    # Research Hub - 嵌入 (使用 OpenAI API，无需本地模型)
    "openai>=1.0.0",
    # LlamaIndex RAG 框架
    "llama-index-core>=0.12.0",
    "llama-index-vector-stores-postgres>=0.4.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-postprocessor-cohere-rerank>=0.3.0",
    "llama-index-readers-file>=0.4.0",
    "llama-index-llms-openai>=0.3.0",
    # Neo4j 图数据库
    "neo4j>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
factor-hub = "domains.factor_hub.cli.main:main"

[project.urls]
Homepage = "https://github.com/quant-research/quant-research-mcp"
Documentation = "https://github.com/quant-research/quant-research-mcp#readme"
Repository = "https://github.com/quant-research/quant-research-mcp"

[tool.setuptools.packages.find]
where = ["backend"]
include = ["app*", "domains*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.txt"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "ARG001", # unused function argument
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG"]

[tool.ruff.lint.isort]
known-first-party = ["factor_hub", "data_hub", "strategy_hub", "mcp_core"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
